refactor(cli): restructure cli directory
This commit refactors the CLI directory structure to better organize the code.

- The main.go file is moved to cmd/tako/.
- The root.go and version.go files are moved to cmd/tako/internal/.
- The package names and import paths are updated accordingly.

This new structure separates the main entry point from the internal command logic.
