+++ cmd/tako/internal/version.go
@@ -1,4 +1,4 @@
-package cmd
+package internal
 
 import (
 	"fmt"
