+++ cmd/tako/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/dangazineu/tako/cmd/tako/internal"
+)
+
+func main() {
+	internal.Execute()
+}
